+++ Expo1900/Expo1900/Controller/ViewController.swift
@@ -10,9 +10,5 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
-
