+++ Expo1900/Expo1900/SupportingFiles/SceneDelegate.swift
@@ -1,0 +1,36 @@
+//
+//  Expo1900 - SceneDelegate.swift
+//  Created by yagom. 
+//  Copyright © yagom academy. All rights reserved.
+// 
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {
+        
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+        
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+        
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        
+    }
+}
